testing appending

+++ include/ft_ssl.h
@@ -6,7 +6,7 @@
 /*   By: bmellon <bmellon@student.42.fr>            +#+  +:+       +#+        */
 /*                                                +#+#+#+#+#+   +#+           */
 /*   Created: 2020/10/20 15:27:52 by bmellon           #+#    #+#             */
-/*   Updated: 2020/10/25 13:58:43 by bmellon          ###   ########.fr       */
+/*   Updated: 2020/10/26 17:19:41 by bmellon          ###   ########.fr       */
 /*                                                                            */
 /* ************************************************************************** */
 
@@ -33,7 +33,7 @@ typedef	struct		s_hash
 {
 	t_hash_flags	flags;
 	size_t			len;
-	int				append;
+	size_t			append;
 	char			*hash;
 	char			*to_hash;
 	char			*type;
